requirements.txt
MouseInfo==0.0.4
Pillow==6.1.0
psutil==5.6.3
PyAutoGUI==0.9.47
PyGetWindow==0.0.6
PyMsgBox==1.0.7
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.22
PyTweening==1.0.3
